@@ -9,29 +9,27 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "14.2.5",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "15.1.9",
+    "react": "19.0.3",
+    "react-dom": "19.0.3",
     "@supabase/supabase-js": "^2.45.0",
-    "@supabase/ssr": "^0.4.0",
-    "next-intl": "^3.17.2",
-    "lucide-react": "^0.424.0",
+    "@supabase/ssr": "^0.5.2",
+    "next-intl": "^3.26.3",
+    "lucide-react": "^0.469.0",
     "clsx": "^2.1.1",
     "tailwind-merge": "^2.4.0",
     "class-variance-authority": "^0.7.0",
-    "react-hook-form": "^7.52.1",
+    "react-hook-form": "^7.54.2",
     "@hookform/resolvers": "^3.9.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "^20.14.12",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@types/node": "^22.10.5",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
     "typescript": "^5.5.4",
     "tailwindcss": "^3.4.7",
     "postcss": "^8.4.40",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "14.2.5"
+    "autoprefixer": "^10.4.19"
   }
 }
